+++ spring-grpc-spring-boot-autoconfigure/src/main/java/org/springframework/grpc/autoconfigure/client/GrpcChannelFactoryConfigurations.java
@@ -23,11 +23,11 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.grpc.autoconfigure.client.GrpcClientAutoConfiguration.NamedChannelVirtualTargets;
 import org.springframework.grpc.client.ChannelCredentialsProvider;
 import org.springframework.grpc.client.ClientInterceptorsConfigurer;
 import org.springframework.grpc.client.GrpcChannelBuilderCustomizer;
 import org.springframework.grpc.client.GrpcChannelFactory;
+import org.springframework.grpc.client.NamedChannelRegistry;
 import org.springframework.grpc.client.NettyGrpcChannelFactory;
 import org.springframework.grpc.client.ShadedNettyGrpcChannelFactory;
 
@@ -47,14 +47,14 @@ class GrpcChannelFactoryConfigurations {
 	static class ShadedNettyChannelFactoryConfiguration {
 
 		@Bean
-		ShadedNettyGrpcChannelFactory shadedNettyGrpcChannelFactory(ChannelBuilderCustomizers channelBuilderCustomizers,
-				ClientInterceptorsConfigurer interceptorsConfigurer, ChannelCredentialsProvider credentials,
-				GrpcClientProperties properties) {
+		ShadedNettyGrpcChannelFactory shadedNettyGrpcChannelFactory(NamedChannelRegistry namedChannelRegistry,
+				ChannelBuilderCustomizers channelBuilderCustomizers,
+				ClientInterceptorsConfigurer interceptorsConfigurer, ChannelCredentialsProvider credentials) {
 			List<GrpcChannelBuilderCustomizer<io.grpc.netty.shaded.io.grpc.netty.NettyChannelBuilder>> builderCustomizers = List
 				.of(channelBuilderCustomizers::customize);
 			var factory = new ShadedNettyGrpcChannelFactory(builderCustomizers, interceptorsConfigurer);
 			factory.setCredentialsProvider(credentials);
-			factory.setVirtualTargets(new NamedChannelVirtualTargets(properties));
+			factory.setVirtualTargets(namedChannelRegistry);
 			return factory;
 		}
 
@@ -67,14 +67,14 @@ class GrpcChannelFactoryConfigurations {
 	static class NettyChannelFactoryConfiguration {
 
 		@Bean
-		NettyGrpcChannelFactory nettyGrpcChannelFactory(ChannelBuilderCustomizers channelBuilderCustomizers,
-				ClientInterceptorsConfigurer interceptorsConfigurer, ChannelCredentialsProvider credentials,
-				GrpcClientProperties properties) {
+		NettyGrpcChannelFactory nettyGrpcChannelFactory(NamedChannelRegistry namedChannelRegistry,
+				ChannelBuilderCustomizers channelBuilderCustomizers,
+				ClientInterceptorsConfigurer interceptorsConfigurer, ChannelCredentialsProvider credentials) {
 			List<GrpcChannelBuilderCustomizer<NettyChannelBuilder>> builderCustomizers = List
 				.of(channelBuilderCustomizers::customize);
 			var factory = new NettyGrpcChannelFactory(builderCustomizers, interceptorsConfigurer);
 			factory.setCredentialsProvider(credentials);
-			factory.setVirtualTargets(new NamedChannelVirtualTargets(properties));
+			factory.setVirtualTargets(namedChannelRegistry);
 			return factory;
 		}
