+++ spring-grpc-spring-boot-autoconfigure/src/test/java/org/springframework/grpc/autoconfigure/client/GrpcClientAutoConfigurationTests.java
@@ -40,10 +40,11 @@ import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
-import org.springframework.grpc.autoconfigure.client.GrpcClientAutoConfiguration.NamedChannelVirtualTargets;
 import org.springframework.grpc.client.ChannelCredentialsProvider;
 import org.springframework.grpc.client.GrpcChannelBuilderCustomizer;
 import org.springframework.grpc.client.GrpcChannelFactory;
+import org.springframework.grpc.client.NamedChannel;
+import org.springframework.grpc.client.NamedChannelRegistry;
 import org.springframework.grpc.client.NettyGrpcChannelFactory;
 import org.springframework.grpc.client.ShadedNettyGrpcChannelFactory;
 
@@ -67,6 +68,26 @@ class GrpcClientAutoConfigurationTests {
 	}
 
 	@Test
+	void whenHasUserDefinedChannelRegistryDoesNotAutoConfigureBean() {
+		NamedChannelRegistry customChannelRegistry = mock(NamedChannelRegistry.class);
+		this.contextRunner()
+			.withBean("customChannelRegistry", NamedChannelRegistry.class, () -> customChannelRegistry)
+			.run((context) -> assertThat(context).getBean(NamedChannelRegistry.class).isSameAs(customChannelRegistry));
+	}
+
+	@Test
+	void channelRegistryAutoConfiguredAsExpected() {
+		this.contextRunner()
+			.run((context) -> assertThat(context).getBean(NamedChannelRegistry.class).satisfies((channelRegistry) -> {
+				var properties = context.getBean(GrpcClientProperties.class);
+				assertThat(channelRegistry.getDefaultChannel()).isEqualTo(properties.getDefaultChannel());
+				assertThat(channelRegistry)
+					.extracting("channels", InstanceOfAssertFactories.map(String.class, NamedChannel.class))
+					.containsExactlyInAnyOrderEntriesOf(properties.getChannels());
+			}));
+	}
+
+	@Test
 	void whenHasUserDefinedCredentialsProviderDoesNotAutoConfigureBean() {
 		ChannelCredentialsProvider customCredentialsProvider = mock(ChannelCredentialsProvider.class);
 		this.contextRunner()
@@ -79,7 +100,7 @@ class GrpcClientAutoConfigurationTests {
 	void credentialsProviderAutoConfiguredAsExpected() {
 		this.contextRunner()
 			.run((context) -> assertThat(context).getBean(NamedChannelCredentialsProvider.class)
-				.hasFieldOrPropertyWithValue("channels", context.getBean(GrpcClientProperties.class))
+				.hasFieldOrPropertyWithValue("channelRegistry", context.getBean(NamedChannelRegistry.class))
 				.extracting("bundles")
 				.isInstanceOf(SslBundles.class));
 	}
@@ -228,7 +249,7 @@ class GrpcClientAutoConfigurationTests {
 				.isInstanceOf(expectedChannelFactoryType)
 				.hasFieldOrPropertyWithValue("credentials", context.getBean(NamedChannelCredentialsProvider.class))
 				.extracting("targets")
-				.isInstanceOf(NamedChannelVirtualTargets.class));
+				.isInstanceOf(NamedChannelRegistry.class));
 	}
 
 	@Test
